Added GPU acceleration for per_camera logic *except for BVLS*.
Modified solve_matrices algorithm to accept PCA or BVLS as a method.
Re-orgianized logic for GPU to take advantage of existing
calc_zchi2_batch algorithm.
For some strange reason refactoring tdata to add rows for color data in
batch in 3d array on CPU takes way longer than a python loop over 2d
arrays so I left two methods for per_camera - one for GPU and one for
CPU - for now while this gets resolved.

This is not the cleanest but it is the fastest and we can revisit if we
change our mind on that.