Added docstring to archetypes:rebin_template_batch
Added always_return_array arg to transmission_Lyman that is True by default.
Changed calls to transmission_Lyman throughout redrock in archetypes, fitz,
templates to call with always_return_array=False to optimize because no need
to generate additional arrays of all 1 and multiply by them.
If always_return_array is True, an array of all ones will be returned in this
case instead of None.