Implemented a CuPy based GPU version of scipy.optimize.lsq_linear's BVLS algorithm.  This is in the optimize folder and will be submitted to CuPy for inclusion at which point it will be removed from Redrock.
zscan has been updated to use this GPU based version instead of our NNLS
trick of adding negative legendre terms and using NNLS to emulate BVLS;
although that trick is preserved via self._solver_method =
"bvls_via_nnls" in archetypes.py where Archetype files are allowed to
specify their preferred solver method (pca, nnls, bvls, or
bvls_via_nnls).

This branch is forked from the branch legendre_nnls_merge_with_main,
which was ready to merge with main and had addressed conflicts that had
arisen during parallel developement by myself and Abhijeet.

zfind is modified slightly to add a timing print statement for the time
spent in each solver method.

Timing Results:
With 4 GPUs and 4 CPUs, BVLS on GPU takes ~17s resulting in a fine
redshift scan time of 72s compared to the NNLS trick taking ~14s for a
fine redshift scan time of 76s (the reason the fitting is faster but the
overall time is slower is due to other operations performed on larger
arrays with the added negative legendre terms).  This compares to using
scipy's CPU-based BVLS, which takes ~63s to do the fitting and 119s for
the fine redshift scan.